feat: extract svg creation in chart component and add tooltip
